web/requirements.txt
Flask
scikit-learn
scikit-image
numpy
joblib
opencv-python
Pillow